# Improve test performance of JujuConnSuite tests by keeping collections.