JujuConnSuite tests are slow due to the creation of collections in mongo being slow.
This PR allows JujuConnSuite to keep collections between each test and only cleanup the database in suite tear down.

This saves around 250ms - 300ms between each test on my local machine.

Before
ok  	github.com/juju/juju/apiserver/facades/client/application	358.573s
After
ok  	github.com/juju/juju/apiserver/facades/client/application	214.015s

Depends on https://github.com/juju/testing/pull/162

## QA steps

Run unit tests.

## Documentation changes

N/A

## Bug reference

N/A